fix: showTreeNode to not expand sibling node
The sibling nodes were sometimes expanded as well as the parent.
