+++ log-viewer/modules/Util.ts
@@ -37,40 +37,6 @@ export function showTab(tabId: string) {
   }
 }
 
-export function showTreeNode(timestamp: number) {
-  const methodElm = document.querySelector(
-      `div[data-enterstamp="${timestamp}"]`
-    ) as HTMLElement,
-    methodName = methodElm?.querySelector("span.name") || methodElm;
-
-  if (methodElm) {
-    showTab("treeTab");
-    expandTreeNode(methodElm, false);
-    methodElm?.scrollIntoView();
-    if (methodName) {
-      window.getSelection()?.selectAllChildren(methodName);
-    }
-  }
-}
-
-function expandTreeNode(elm: HTMLElement, expand: boolean) {
-  if (expand) {
-    const toggle = elm.querySelector(".toggle"),
-      childContainer = elm.querySelector(".childContainer") as HTMLElement;
-
-    if (childContainer && toggle) {
-      childContainer.style.display = "block";
-      toggle.textContent = "-";
-    }
-  }
-
-  const parent = elm.parentElement;
-  if (parent && parent.id !== "tree") {
-    // stop at the root of the tree
-    expandTreeNode(parent, true);
-  }
-}
-
 export function recalculateDurations(node: LogLine) {
   if (node.exitStamp) {
     node.netDuration = node.duration = node.exitStamp - node.timestamp;
